% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{color_correction}
\alias{color_correction}
\title{Match color between two images}
\usage{
color_correction(image1, image2, blur = 0.6)
}
\arguments{
\item{image1}{Reference image}

\item{image2}{Image to color correct}
}
\value{

}
\description{
Match color between two images
}
\examples{

}
