% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_align.R
\name{face_align}
\alias{face_align}
\title{Align two images}
\usage{
face_align(target, reference, scope = c("linear", "nonlinear"), save = TRUE,
  ...)
}
\arguments{
\item{target}{Path to target image}

\item{reference}{Path to reference image (what to alignt to)}

\item{scope}{Perform a 'linear' or 'nonlinear' transformation (see \link[RNiftyReg]{niftyreg})}

\item{save}{Logical indicating if the aligned image should be saved. Defaults to \code{TRUE}}

\item{...}{Optional arguments to pass to \link[RNiftyReg]{niftyreg}}
}
\description{
Align two images
}
