% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_coord.R
\name{face_coord}
\alias{face_coord}
\title{Face Coordinates}
\usage{
face_coord(python_location = "/usr/local/bin/python", facedetector = NULL,
  image)
}
\arguments{
\item{python_location}{The location of your python install. Defaults to
default install location for MacOS (e.g., \code{'/usr/local/bin/python'}).}

\item{facedetector}{External python script location for obtaining bounding box.
By default, it looks for this in the package's \code{data/} folder, but can
also manually specify.}

\item{image}{A string specifying the input image to detect face(s) in.}
}
\value{
Returns number of faces by 4 dimension dataframe specifying the bounding box
coordinates of each face and the width and height of each bonding box
}
\description{
Face Coordinates
}
\examples{
im <- system.file('data','faces.jpg',package = 'quantIm')
face_coord(image=im)

...
}
\author{
Daniel N. Albohn
}
