% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_crop.R
\name{face_crop}
\alias{face_crop}
\title{Crop image to face}
\usage{
face_crop(coords, image, savename = NULL)
}
\arguments{
\item{coords}{Face coordinates provided by \code{\link{face_coord}}}

\item{image}{Image name to be cropped}

\item{savename}{File name to save cropped image. Defaults to 'cropped' + \code{image}.png.}
}
\value{
Saves cropped image.
}
\description{
Crop image to face
}
\examples{
image <- system.file('data','faces.jpg',package = 'quantIm')
coords <- face_coord(image)
face_crop(coords,image)
...
}
