% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_crop_points_III.R
\name{face_crop_points}
\alias{face_crop_points}
\alias{Daniel}
\alias{N.}
\alias{Albohn}
\title{Crop image around set of points}
\usage{
face_crop_points(coords = NULL, image, points = "default",
  savename = NULL, eye_dist = 125, rescale = TRUE, wh = 512, ...)
}
\arguments{
\item{dist}{}
}
\description{
Crop image around set of points
}
