% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\name{face_landmarks}
\alias{face_landmarks}
\title{Detect Face Landmarks and Bounding Box}
\usage{
face_landmarks(image, python = NULL, condaenv = NULL)
}
\arguments{
\item{image}{A string specifying the input image to detect face(s) in.}

\item{python}{}
}
\value{
Returns a list which contains dlib's 68 facial landmarks
as well as the bounding box that the face in the image exists.
}
\description{
A simple wrapper for dlib's face detector algorithm.
}
\details{
The following python modules must be installed and accessible to the version of python
specified in \code{python_location} (see parameters):
\itemize{
    \item{\code{imutils}}
    \item{\code{cv2}}
    \item{\code{dlib}}
    \item{\code{numpy}}
}
All of these can be installed from the command line with \code{pip install} on Unix systems
except for \code{openCV2} which needs to be downloaded. Additionally, the following
packages must be installed and up to date:

\itemize{
    \item{\code{sudo apt-get install cmake}}
   \item{\code{sudo apt-get install libboost-all-dev}}
   \item{\code{sudo apt-get install libmagick++-dev}}
   \item{\code{sudo apt-get install libopencv-dev python-opencv}}
}
}
\examples{
image = system.file("extdata", "defiant2.jpg", package = "quantIm")
face_landmarks(image=image, python=NULL)
----
}
\seealso{
dlib \link{http://dlib.net/face_landmark_detection.py.html}
OpenCV2 \link{http://opencv.org/platforms/}
imagemagick and magick \link{https://ropensci.org/blog/2016/08/23/z-magick-release}
}
\author{
Daniel N. Albohn
}
