% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_python.R
\name{find_python}
\alias{find_python}
\title{Helper function to find python on user system}
\usage{
find_python(python)
}
\arguments{
\item{python}{}
}
\description{
Helper function to find python on user system
}
