% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{get_landmarks}
\alias{get_landmarks}
\title{Get facial landmarks}
\usage{
get_landmarks(image, convert = TRUE)
}
\arguments{
\item{image}{Path to image}

\item{convert}{Convert output to be R compatible?}
}
\value{
Dataframe with 68 face points.
}
\description{
Get facial landmarks
}
\examples{
img <- system.file("extdata", "obama.png", package = "quantIm")
get_landmarks(img)

}
