% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_heatmap.R
\name{im_heatmap}
\alias{im_heatmap}
\title{Create heatmap image from t-values}
\usage{
im_heatmap(path = NULL, name, file = NULL, base_image, mask = NULL,
  thresh = 1.96)
}
\arguments{
\item{path}{Absolute path to t-values file from working directory}

\item{file}{Name of actual image}

\item{base_image}{Relative path to base image file (in \code{.jpg} or \code{.png} format).}

\item{mask}{Relative path to mask image file (in \code{.jpg} or \code{.png} format). Defaults
to \code{NULL}.}

\item{thresh}{}
}
\description{
Create heatmap image from t-values
}
