% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsub.R
\name{im_sub}
\alias{im_sub}
\title{Subtract two images}
\usage{
im_sub(image1, image2, folder1 = NULL, folder2 = NULL, path = NULL,
  resize = FALSE, write.data = TRUE, sub = "default", blur = NULL)
}
\arguments{
\item{blur}{}
}
\description{
Subtract two images
}
