% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_image_functions.R
\name{im_t_test}
\alias{im_t_test}
\title{Compute t-test on multiple images}
\usage{
im_t_test(path, pattern1 = "*.csv", name, write = TRUE)
}
\arguments{
\item{path}{File path to folder with subtracted image files and values.}

\item{pattern1}{Pattern that the subtracted files were saved in. Defaults to '.csv'}

\item{name}{Name that the t-values file should be saved as.}

\item{write}{Should the file be saved (\code{TRUE}), or returned as a vector (\code{FALSE}).
Defaults to \code{TRUE}.}
}
\value{

}
\description{
Compute t-test on multiple images
}
