% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_image.R
\name{pad_image}
\alias{pad_image}
\title{Pads an image to a certain width and height}
\usage{
pad_image(image, width, height, background, out)
}
\arguments{
\item{background}{}
}
\description{
Pads an image to a certain width and height
}
