% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_face_functions.R
\name{transform_points}
\alias{transform_points}
\title{Create affine transformation}
\usage{
transform_points(landmarks1, landmarks2)
}
\arguments{
\item{landmarks2}{}
}
\description{
Create affine transformation
}
\examples{

}
