% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{warp_image}
\alias{warp_image}
\title{Warp (affine transform) an image}
\usage{
warp_image(image, m, width, height, convert = FALSE)
}
\arguments{
\item{m}{Affine matrix.}

\item{convert}{Convert output to be R compatible?}

\item{image1}{Path to image to warp.}
}
\value{
Warped image
}
\description{
Warp (affine transform) an image
}
\examples{

}
