% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_crop.R
\name{batch_crop}
\alias{batch_crop}
\title{Crop Batches of Images from a Folder}
\usage{
batch_crop(file_list, path, save_string, dist, wh, points, gleam = FALSE,
  pad = FALSE)
}
\arguments{
\item{pad}{}
}
\description{
Crop Batches of Images from a Folder
}
\examples{

}
