% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_image_functions.R
\name{batch_im_sub}
\alias{batch_im_sub}
\title{Subtract batches of images}
\usage{
batch_im_sub(files1, files2, save_image, write_data, ...)
}
\arguments{
\item{blur}{}
}
\value{

}
\description{
Subtract batches of images
}
\examples{

}
