% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{distance_metrics}
\alias{distance_metrics}
\title{Obtain Various Distance Metrics Between Two Images}
\usage{
distance_metrics(target, reference, save_id = NULL, to_file = TRUE,
  python = NULL)
}
\arguments{
\item{target}{Path to an image.}

\item{reference}{Path to an image to be used as a reference}

\item{save_id}{Regular expression to extract ID from file name. Defaults
to full path provided for file if \code{NULL}.}

\item{to_file}{Logical. Should results be written to file or printed in console?}

\item{python}{Which installation of python should be used?}
}
\value{
Returns dataframe with various distance metrics. See details for
which distance measures are includes.
}
\description{
Obtain Various Distance Metrics Between Two Images
}
\examples{

}
