% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_metrics.R
\name{distance_metrics}
\alias{distance_metrics}
\title{Obtain Various Distance Metrics Between Two Images}
\usage{
distance_metrics(target, reference, save_id = NULL)
}
\arguments{
\item{target}{Path to an image.}

\item{reference}{Path to an image to be used as a reference}

\item{save_id}{Regular expression to extract ID from file name. Defaults
to full path provided for file if \code{NULL}.}
}
\value{
Returns dataframe with various distance metrics. See details for
which distance measures are includes.
}
\description{
Obtain Various Distance Metrics Between Two Images
}
\details{
The following distance metrics are computed:
nrmse= Normalized root mean squred error from python \code{skimage.measure}
module (must be installed prior to use),
psnr = Peak Signal-to-Noise ration from python \code{skimage.measure} module (must be installed priot to use),
ssim = Structural similarity index from python \code{skimage.measure} module (must be installed prior to use),
sim_reg = Similarity metric from \code{\link{RNiftyReg::similarity}},
dtw = Dynamic Time Warping distance from \code{\link{dtwclust::dtw_lb}},
fdtw = NULL,
\link{https://pypi.python.org/pypi/fastdtw} (must be installed prior to use),
fourier = Fast Fourier Transform distance from \code{\link{TSdist::FourierDistance}},
emd = Earth Mover's Distance computed from
\link{https://github.com/garydoranjr/pyemd} (must be installed prior to use),
dissim = NULL.

All of the required python packages used for this function can be installed by running the
setup function \code{quantIm::initialize_quantIm()}. Note that this setup function only works for
UNIX based operating systems (not Windows).
}
\examples{

}
