% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_crop.R
\name{face_crop}
\alias{face_crop}
\title{Crop an image to the face}
\usage{
face_crop(coords, image, savename = NULL)
}
\arguments{
\item{coords}{Face coordinates provided by \code{\link{face_coord}}}

\item{image}{Image name to be cropped}

\item{savename}{File name to save cropped image. Defaults to "cropped_\code{image}.png."}
}
\value{
Saves cropped image as a \code{.png} image file.
}
\description{
Crop an image to the face
}
\examples{
image <- system.file('data','faces.jpg',package = 'quantIm')
coords <- face_coord(image)
face_crop(coords,image)
...
}
\author{
Daniel N. Albohn
}
