% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_crop_points.R
\name{face_crop_points}
\alias{face_crop_points}
\alias{Daniel}
\alias{N.}
\alias{Albohn}
\title{Crop image around set of points}
\usage{
face_crop_points(coords = NULL, image, points = c(30, 28),
  savename = NULL, dist = "116", wh = "512")
}
\arguments{
\item{dist}{}
}
\description{
Crop image around set of points
}
