% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_crop_points.R
\name{face_crop_points}
\alias{face_crop_points}
\alias{Daniel}
\alias{N.}
\alias{Albohn}
\title{Crop image around set of points}
\usage{
face_crop_points(coords, image, points = c(30, 28), savename = NULL,
  scale = 1, wh = NULL)
}
\arguments{
\item{wh}{}
}
\description{
Crop image around set of points
}
