% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_heatmap.R
\name{im_heatmap}
\alias{im_heatmap}
\title{Create heatmap image from t-values}
\usage{
im_heatmap(path, file = paste0(name, "_average_t_values.csv"), base_image,
  mask = NULL, thresh = 1.96)
}
\arguments{
\item{thresh}{}
}
\description{
Create heatmap image from t-values
}
