% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_heatmap.R
\name{im_heatmap}
\alias{im_heatmap}
\title{Create heatmap image from t-values}
\usage{
im_heatmap(tz_file, base_image, mask, thresh = 1.96, absval = FALSE)
}
\arguments{
\item{tz_file}{Path to file of t-values.}

\item{base_image}{Relative path to base image file (in \code{.jpg} or \code{.png} format).}

\item{mask}{Relative path to mask image file (in \code{.jpg} or \code{.png} format). Defaults
to \code{NULL}.}

\item{thresh}{The cuttoff score for significance. Defaults to p = .05, or 1.96.}

\item{absval}{Logical. Should the absolute value of the z/t scores be taken?}
}
\value{
Returns a \code{ggplot2} object that can be further manipulated using
\code{ggplot2} functions, or to be saved.
}
\description{
Create heatmap image from t-values
}
\details{
If saving graph, make sure to save the image with the exact width and
height of the original images, else the heatmap values will not be placed correctly.
}
