% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imsub.R
\name{im_sub}
\alias{im_sub}
\title{Subtract two images}
\usage{
im_sub(image1, image2, resize = FALSE, write.data = TRUE, sub = "default",
  blur = 0.3)
}
\arguments{
\item{blur}{}
}
\description{
Subtract two images
}
