% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_image_functions.R
\name{im_sub}
\alias{im_sub}
\title{Subtract two images}
\usage{
im_sub(image1, image2, save_image = TRUE, write_data = TRUE, blur)
}
\arguments{
\item{write_data}{}
}
\description{
Subtract two images
}
