% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im_t_test.R
\name{im_t_test}
\alias{im_t_test}
\title{Compute t-test on multiple images}
\usage{
im_t_test(path = NULL, pattern1 = "*.csv", name)
}
\arguments{
\item{name}{}
}
\description{
Compute t-test on multiple images
}
