% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_of.R
\name{python_of}
\alias{python_of}
\title{Compute optical flow between two images}
\usage{
python_of(reference, target, output, python = NULL)
}
\arguments{
\item{python}{}
}
\description{
Compute optical flow between two images
}
