% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_face_functions.R
\name{transform_points}
\alias{transform_points}
\title{Affine transform face image points}
\usage{
transform_points(image1, image2, convert = TRUE)
}
\arguments{
\item{image1}{Image to reference in transformation.}

\item{image2}{Image to be transformed.}

\item{convert}{Convert output to be R compatible?}
}
\value{
Affine transformation matrix.
}
\description{
Affine transform face image points
}
\examples{

}
